Updated concat.js
Added support for processing file contents. User will now be able to provide a function as part of config options which will be used to process file contents before being appended. The function will be provided file content and file path as inputs in the order mentioned and will be expected to return the possibly altered file content.

Added support to inform the user of missing files. Missing files will now be logged to the console.